+++ app/src/main/java/ch/epfl/sdp/cook4me/Cook4MeApp.kt
@@ -7,12 +7,15 @@ import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
 import ch.epfl.sdp.cook4me.ui.LoginScreen
 import ch.epfl.sdp.cook4me.ui.SwipeScreen
+import ch.epfl.sdp.cook4me.ui.TupCreationScreenWithState
+import ch.epfl.sdp.cook4me.ui.TupCreationViewModel
 
 /**
  * enum values that represent the screens in the app
  */
 private enum class Screen {
     Login,
+    TupCreationScreen,
     SwipeScreen
 }
 
@@ -25,6 +28,9 @@ fun Cook4MeApp(
             LoginScreen(
                 onSuccessfulLogin = { navController.navigate(Screen.SwipeScreen.name) }
             )
+        }
+        composable(route = Screen.TupCreationScreen.name) {
+            TupCreationScreenWithState(TupCreationViewModel())
         }
         composable(route = Screen.SwipeScreen.name) {
             SwipeScreen()
