+++ app/src/main/java/ch/epfl/sdp/cook4me/application/TupperwareServiceWithRepository.kt
@@ -3,7 +3,8 @@ package ch.epfl.sdp.cook4me.application
 import ch.epfl.sdp.cook4me.persistence.model.Tupperware
 import ch.epfl.sdp.cook4me.persistence.repository.TupperwareRepository
 
-class TupperwareServiceWithRepository(private val repository: TupperwareRepository) : TupperwareService {
+class TupperwareServiceWithRepository(private val repository: TupperwareRepository = TupperwareRepository()) :
+    TupperwareService {
 
     override suspend fun submitForm(
         title: String,
